abundance search [more fixes]
This commit addresses an issue in the explorer search database function, which was returning datasets without respecting user-defined abundance threshold parameters. This commit also accounts for situations in which the sum of values attempts to be divided by zero in generating the average abundance value for a site, which would result in a server error, failed api call, and unresponsive explorer application.